Better formatting for times and added paper dispositions
The disposition stuff is commented out for now until I can improve the formatting.  I want
to use a dropdown for selecting the disposition of each paper and then (hopefully uniformly
spaced) badges (using a table?) at the top level if the user has expressed dispositions
on a per paper basis.